lightrec: Don't allocate data buffer if disassembler is disabled
The data buffer in Lightning is used for storing constants (which are
only used for the float and double operations that have an immediate
argument, but Lightrec doesn't do float/double) or storing notes.

If the disassembler is disabled, there is no need for Lightning to
allocate the data buffer, so we can request it to skip its creation.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>